Fix doc and type in image content header

+++ done/pictDBM/image_content.h
@@ -11,10 +11,11 @@
  * @brief Resizes an image to a new resolution.
  *
  * @param resolution The resolution of the image to create. Possible values are
- * RES_THUMB, RES_SMALL, RES_ORIG
- * @param db_file The database file containing the image to resize and the resized
- * image.
- * @param index The index of the image to resize.
+ *                   RES_THUMB, RES_SMALL, RES_ORIG.
+ * @param db_file    The database file containing the image to resize
+ *                   and the resized image.
+ * @param index      The index of the image to resize in the database.
+ * @return 0 if the resize was successful, a non-zero int otherwise.
  */
-int lazily_resize(uint16_t resolution, struct pictdb_file* db_file,
+int lazily_resize(int resolution, struct pictdb_file* db_file,
                   size_t index);
